Repair the BUG in kata2.uvprojx:
Somehow this configuration changed so it ended up using a source file
from another project (serial-io.cpp).  This was hard to notice at
first, but it meant that changes to that file were not actually
changes to the correct file, and so the build was broken and I
committed the code regardless.

Since I already pushed the broken version 5bb9a45, I won't actually
repair that and cause grief to the students.  Instead I've manuall
inserted a correcting commit (9fd4a9a) and I'll rebase the rest on top
of that.